+++ CHANGES.md
@@ -28,6 +28,7 @@ Bugfix 🐛:
  - Invalid popup when pressing back (#1635)
  - Simplifies draft management and should fix bunch of draft issues (#952, #683)
  - Very long topic cannot be fully visible (#1957)
+ - Properly detect cross signing keys reset
 
 Translations 🗣:
  - Move store data to `/fastlane/metadata/android` (#812)